feat: simplify schema for ByteArray
We used Uint8Array but it doesn't have a nice JSON serialization and
also causing a lot of Hex string <-> Uint8Array conversion. So I decided
to go back to Hex (a type alias of string) - which is less typesafe but
acceptable since most contracts' types are ByteArray, and String are
quite rare.